requirements.txt
fastapi==0.116.1
uvicorn==0.34.0
python-jose==3.3.0
python-multipart==0.0.17
pydantic==2.10.4
pydantic-settings==2.7.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0 # Для загрузки .env